% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preprocess.R
\name{d4h_clean_band}
\alias{d4h_clean_band}
\title{Clean Raster Band Artefacts}
\usage{
d4h_clean_band(raster_layer, min_val = -1, max_val = 1)
}
\arguments{
\item{raster_layer}{SpatRaster. The input band to clean.}

\item{min_val}{Numeric. Minimum valid value.}

\item{max_val}{Numeric. Maximum valid value.}
}
\value{
A cleaned SpatRaster.
}
\description{
Replaces NaN and Infinite values with NA and clamps the raster to physically valid limits.
}
