% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_ifeb}
\alias{d4h_ifeb}
\title{Fragmentation and Edge Effect Index (IFEB)}
\usage{
d4h_ifeb(grad_red_edge, red, nir, eps = 0.001)
}
\arguments{
\item{grad_red_edge}{SpatRaster or character. Focal standard deviation of Red Edge or path to file.}

\item{red}{SpatRaster or character. Red band or path to file.}

\item{nir}{SpatRaster or character. Near-infrared band or path to file.}

\item{eps}{Numeric. Small epsilon to prevent division by zero (default: 0.001).}
}
\value{
A SpatRaster quantifying edge effects with background masked to \code{NA}.
}
\description{
Quantifies ecological transitions and ecotone borders between tree canopies and urban or cleared soil substrates.
}
\details{
The Forest Edge & Environmental Barrier Index is formulated as:
\deqn{\text{IFEB} = |\nabla(\text{RedEdge})| \times \frac{\text{Red}}{\text{NIR} + \epsilon}}
}
\references{
Laurance, W. F., et al. (2011). The impact of forest fragmentation on disease ecology. \emph{Biological Conservation}, 144(1), 56-68.
}
