% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_imsr_general}
\alias{d4h_imsr_general}
\title{Continuous Synthetic Materials Risk Index (General Area)}
\usage{
d4h_imsr_general(
  red_edge,
  red,
  nir,
  veg_threshold = 0.2,
  eps = 0.001,
  scale = NULL
)
}
\arguments{
\item{red_edge}{Character or SpatRaster. Red Edge band or path to file.}

\item{red}{Character or SpatRaster. Red band or path to file.}

\item{nir}{Character or SpatRaster. Near-infrared band or path to file.}

\item{veg_threshold}{Numeric. Threshold to differentiate vegetation from synthetics (default: 0.2).}

\item{eps}{Numeric. Small epsilon to prevent division by zero (default: 0.001).}

\item{scale}{Numeric. Divisor to normalize pixel values (default: NULL).}
}
\value{
A SpatRaster containing continuous IMSR risk values.
}
\description{
Detects non-natural materials (plastics, rubber, tires) that serve as potential mosquito breeding containers.
}
