% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{d4h_list_metrics}
\alias{d4h_list_metrics}
\title{List Available Micro-Environmental and Epidemiological Metrics}
\usage{
d4h_list_metrics()
}
\value{
A tibble with metric descriptions.
}
\description{
Returns a tibble listing the curated drone metrics, resolutions, and typical applications.
}
