% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ndre}
\alias{d4h_ndre}
\title{Normalized Difference Red Edge Index (NDRE)}
\usage{
d4h_ndre(nir, red_edge)
}
\arguments{
\item{nir}{SpatRaster or character. Near-infrared band or path to file.}

\item{red_edge}{SpatRaster or character. Red Edge band or path to file.}
}
\value{
A SpatRaster containing NDRE values in \code{[-1, 1]} with background masked to \code{NA}.
}
\description{
Calculates the Normalized Difference Red Edge Index (NDRE) using the Red Edge band to evaluate canopy chlorophyll content and late-stage vegetative health.
}
\details{
The Normalized Difference Red Edge Index is formulated as:
\deqn{\text{NDRE} = \frac{\text{NIR} - \text{RedEdge}}{\text{NIR} + \text{RedEdge}}}
Useful in dense vegetation where standard NDVI experiences saturation.
}
\references{
Barnes, E. M., Clarke, T. R., Richards, S. E., Colaizzi, P. D., Sloan, J., Moran, M. S., ... & Pinter, P. J. (2000). Coincident detection of crop water stress, nitrogen status and canopy density using ground-based multispectral data. \emph{Proceedings of the 5th International Conference on Precision Agriculture}, 16, 1-15.
}
