% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grids.R
\name{d4h_raster_general}
\alias{d4h_raster_general}
\title{Convert Raster Indicator to General Homogeneous Mosaic}
\usage{
d4h_raster_general(raster_in)
}
\arguments{
\item{raster_in}{SpatRaster or character. Input raster layer, stack, or path to file.}
}
\value{
A SpatRaster with the mosaic footprint filled with the global mean value.
}
\description{
Creates a continuous SpatRaster matching the exact spatial footprint of the input raster where all valid pixels are assigned the overall global mean value.
}
