% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{d4h_savi_general}
\alias{d4h_savi_general}
\title{Cálculo general continuo de SAVI para toda el área (sin grilla)}
\usage{
d4h_savi_general(nir, red, L = 0.5, escala = NULL)
}
\description{
Cálculo general continuo de SAVI para toda el área (sin grilla)
}
