% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_terrain.R
\name{d4h_slope}
\alias{d4h_slope}
\title{Topographic Slope}
\usage{
d4h_slope(dem_raster, unit = "degrees")
}
\arguments{
\item{dem_raster}{SpatRaster or character. Input DEM/DSM or path to file.}

\item{unit}{Character. Slope unit: "degrees" or "radians" (default: "degrees").}
}
\value{
A SpatRaster containing slope values with background masked to \code{NA}.
}
\description{
Derives micro-terrain slope in degrees or radians from a high-resolution Digital Elevation Model (DEM) or Digital Surface Model (DSM).
}
\details{
Topographic slope represents the rate of maximum elevation change:
\deqn{\beta = \arctan\left(\sqrt{\left(\frac{\partial z}{\partial x}\right)^2 + \left(\frac{\partial z}{\partial y}\right)^2}\right)}
Flat terrains (\eqn{\beta < 5^\circ}) are critical in spatial epidemiology as they favor persistent micro-pooling and water stagnation.
}
\references{
Burrough, P. A., & McDonnell, R. A. (1998). \emph{Principles of Geographical Information Systems}. Oxford University Press.
}
