% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_terrain.R
\name{d4h_twi_general}
\alias{d4h_twi_general}
\title{Continuous Topographic Wetness Index (General Area)}
\usage{
d4h_twi_general(dem, eps = 0.001)
}
\arguments{
\item{dem}{Character or SpatRaster. Digital Elevation/Surface Model (DEM/DSM) or path to file.}

\item{eps}{Numeric. Small epsilon to prevent division by zero (default: 0.001).}
}
\value{
A SpatRaster containing continuous TWI values.
}
\description{
Calculates continuous TWI from a Digital Elevation Model (DEM) to locate water accumulation zones.
}
