% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procesar_mosaico.R
\name{procesar_mosaico_multivariado}
\alias{procesar_mosaico_multivariado}
\title{Pipeline multivariado: Cálculo conjunto de NDVI y SAVI, grilla zonal y exportación}
\usage{
procesar_mosaico_multivariado(
  nir,
  red,
  tamano_grilla = 50,
  square = FALSE,
  out_tif = NULL,
  out_csv = NULL
)
}
\description{
Pipeline multivariado: Cálculo conjunto de NDVI y SAVI, grilla zonal y exportación
}
