% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grillas.R
\name{resumir_por_grilla}
\alias{resumir_por_grilla}
\title{Resumir indicador por grilla}
\usage{
resumir_por_grilla(
  raster_in,
  tamano_celda = 20,
  square = TRUE,
  descartar_na = TRUE
)
}
\description{
Resumir indicador por grilla
}
