% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ctvi}
\alias{d4h_ctvi}
\title{Corrected Transformed Vegetation Index (CTVI)}
\usage{
d4h_ctvi(ndvi_raster)
}
\arguments{
\item{ndvi_raster}{SpatRaster. NDVI input layer.}
}
\value{
A SpatRaster containing CTVI values.
}
\description{
Calculates CTVI to normalize vegetation distributions.
}
