% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ctvi}
\alias{d4h_ctvi}
\title{Corrected Transformed Vegetation Index (CTVI)}
\usage{
d4h_ctvi(ndvi_raster)
}
\arguments{
\item{ndvi_raster}{SpatRaster or character. Input NDVI layer or path to file.}
}
\value{
A SpatRaster containing CTVI values with background masked to NA.
}
\description{
Calculates CTVI to normalize vegetation index distributions.
}
