% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_dvi}
\alias{d4h_dvi}
\title{Difference Vegetation Index (DVI)}
\usage{
d4h_dvi(nir, red)
}
\arguments{
\item{nir}{SpatRaster or character. Near-infrared band or path to file.}

\item{red}{SpatRaster or character. Red band or path to file.}
}
\value{
A SpatRaster containing DVI values with background masked to NA.
}
\description{
Calculates DVI as the simple difference between NIR and Red reflectance.
}
