% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_dvi}
\alias{d4h_dvi}
\title{Difference Vegetation Index (DVI)}
\usage{
d4h_dvi(nir, red)
}
\arguments{
\item{nir}{SpatRaster or character. Near-infrared band.}

\item{red}{SpatRaster or character. Red band.}
}
\value{
A SpatRaster containing DVI values.
}
\description{
Calculates DVI with automatic band alignment.
}
