% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_evi}
\alias{d4h_evi}
\title{Enhanced Vegetation Index (EVI)}
\usage{
d4h_evi(nir, red, blue, g_factor = 2.5, c1 = 6, c2 = 7.5, l_factor = 1)
}
\arguments{
\item{nir}{SpatRaster. Near-infrared band.}

\item{red}{SpatRaster. Red band.}

\item{blue}{SpatRaster. Blue band.}

\item{g_factor}{Numeric. Gain factor.}

\item{c1}{Numeric. Aerosol resistance coefficient 1.}

\item{c2}{Numeric. Aerosol resistance coefficient 2.}

\item{l_factor}{Numeric. Canopy background adjustment.}
}
\value{
A SpatRaster containing EVI values.
}
\description{
Calculates EVI for areas with dense canopy background.
}
