% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_gndvi}
\alias{d4h_gndvi}
\title{Green Normalized Difference Vegetation Index (GNDVI)}
\usage{
d4h_gndvi(nir, green)
}
\arguments{
\item{nir}{SpatRaster or character. Near-infrared band or path to file.}

\item{green}{SpatRaster or character. Green band or path to file.}
}
\value{
A SpatRaster containing GNDVI values in \code{[-1, 1]} with background masked to NA.
}
\description{
Calculates GNDVI for estimating chlorophyll concentration.
}
