% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_iead}
\alias{d4h_iead}
\title{Clearing Stagnation Index (IEAD)}
\usage{
d4h_iead(grad_delta_ndvi, depressions_dsm, red_edge)
}
\arguments{
\item{grad_delta_ndvi}{SpatRaster. Gradient of NDVI change.}

\item{depressions_dsm}{SpatRaster. Topographic depressions from DSM.}

\item{red_edge}{SpatRaster. Red Edge band.}
}
\value{
A SpatRaster highlighting stagnation in clearings.
}
\description{
Detects puddles in recently cleared or deforested areas.
}
