% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_iead}
\alias{d4h_iead}
\title{Clearing Stagnation Index (IEAD)}
\usage{
d4h_iead(grad_delta_ndvi, depressions_dsm, red_edge)
}
\arguments{
\item{grad_delta_ndvi}{SpatRaster or character. Gradient of NDVI change layer or path to file.}

\item{depressions_dsm}{SpatRaster or character. Topographic depressions layer from DSM or path to file.}

\item{red_edge}{SpatRaster or character. Red Edge band or path to file.}
}
\value{
A SpatRaster highlighting stagnation risk in clearings with background masked to NA.
}
\description{
Detects surface water pooling in recently cleared or deforested areas.
}
