% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_iec}
\alias{d4h_iec}
\title{Breeding Site Stratification Index (IEC)}
\usage{
d4h_iec(gndvi_water, lst, t_opt = 25)
}
\arguments{
\item{gndvi_water}{SpatRaster or character. GNDVI layer masked to water bodies or path to file.}

\item{lst}{SpatRaster or character. Land Surface Temperature layer in Celsius or path to file.}

\item{t_opt}{Numeric. Optimal incubation temperature in degrees Celsius (default: 25).}
}
\value{
A SpatRaster stratifying larval breeding site quality with background masked to NA.
}
\description{
Evaluates open water bodies weighted by optimal thermal ranges for larval development.
}
