% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_iec}
\alias{d4h_iec}
\title{Breeding Site Stratification Index (IEC)}
\usage{
d4h_iec(gndvi_water, lst, t_opt = 25)
}
\arguments{
\item{gndvi_water}{SpatRaster. GNDVI masked strictly to water bodies.}

\item{lst}{SpatRaster. Land Surface Temperature in Celsius.}

\item{t_opt}{Numeric. Optimal incubation temperature.}
}
\value{
A SpatRaster stratifying breeding site quality.
}
\description{
Evaluates water bodies weighted by optimal thermal ranges for larval development.
}
