% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_iev}
\alias{d4h_iev}
\title{Vulnerability Stagnation Indicator (IEV)}
\usage{
d4h_iev(delta_ndwi, slope_dsm, eps = 0.001)
}
\arguments{
\item{delta_ndwi}{SpatRaster. Temporal difference in NDWI or base NDWI.}

\item{slope_dsm}{SpatRaster. Slope derived from the Digital Surface Model.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster highlighting stagnation risk.
}
\description{
Locates areas where surface water is likely to stagnate due to micro-depressions.
}
