% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_ifeb}
\alias{d4h_ifeb}
\title{Fragmentation and Edge Effect Index (IFEB)}
\usage{
d4h_ifeb(grad_red_edge, red, nir, eps = 0.001)
}
\arguments{
\item{grad_red_edge}{SpatRaster. Focal standard deviation of the Red Edge band.}

\item{red}{SpatRaster. Red band.}

\item{nir}{SpatRaster. Near-infrared band.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster quantifying edge effects.
}
\description{
Quantifies ecological transitions and borders between vegetation and urban/cleared soil.
}
