% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_imsr}
\alias{d4h_imsr}
\title{Synthetic Materials Risk Index (IMSR)}
\usage{
d4h_imsr(red_edge, red, nir, veg_threshold = 0.2, eps = 0.001)
}
\arguments{
\item{red_edge}{SpatRaster. Red Edge band.}

\item{red}{SpatRaster. Red band.}

\item{nir}{SpatRaster. Near-infrared band.}

\item{veg_threshold}{Numeric. Threshold to differentiate vegetation from synthetics.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster highlighting synthetic materials.
}
\description{
Detects non-natural materials (plastics, rubber, tires) that can serve as vector breeding sites.
}
