% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_irhe}
\alias{d4h_irhe}
\title{Spectral Water Retention Index (IRHE)}
\usage{
d4h_irhe(green, nir, lst, eps = 0.001)
}
\arguments{
\item{green}{SpatRaster. Green band.}

\item{nir}{SpatRaster. Near-infrared band.}

\item{lst}{SpatRaster. Land Surface Temperature.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster representing water retention potential.
}
\description{
Highlights stable, cold water bodies with low evaporation rates.
}
