% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_irih}
\alias{d4h_irih}
\title{Human Interface Roughness Index (IRIH)}
\usage{
d4h_irih(var_re_nir, red, green, eps = 0.001)
}
\arguments{
\item{var_re_nir}{SpatRaster. Local variance of the RE/NIR ratio.}

\item{red}{SpatRaster. Red band.}

\item{green}{SpatRaster. Green band.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster representing structural roughness.
}
\description{
Measures canopy structural heterogeneity in the forest-household interface.
}
