% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_epidemiology.R
\name{d4h_irts}
\alias{d4h_irts}
\title{Thermal Refuge and Shade Index (IRTS)}
\usage{
d4h_irts(ndre, lst_local, lst_surrounding, nir, eps = 0.001)
}
\arguments{
\item{ndre}{SpatRaster. NDRE index layer.}

\item{lst_local}{SpatRaster. Local Land Surface Temperature.}

\item{lst_surrounding}{SpatRaster. Focal mean surrounding Land Surface Temperature.}

\item{nir}{SpatRaster. Near-infrared band.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster representing thermal refuges.
}
\description{
Identifies cool, shaded microhabitats serving as vector refuges.
}
