% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_ml.R
\name{d4h_list_metrics}
\alias{d4h_list_metrics}
\title{List Available Micro-Environmental and Epidemiological Metrics}
\usage{
d4h_list_metrics()
}
\value{
A tibble with metric metadata.
}
\description{
Returns a tibble listing available metrics, resolutions, and typical applications.
}
