% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ndre}
\alias{d4h_ndre}
\title{Normalized Difference Red Edge Index (NDRE)}
\usage{
d4h_ndre(nir, red_edge)
}
\arguments{
\item{nir}{SpatRaster or character. Near-infrared band or path to file.}

\item{red_edge}{SpatRaster or character. Red Edge band or path to file.}
}
\value{
A SpatRaster containing NDRE values in \code{[-1, 1]} with background masked to NA.
}
\description{
Calculates NDRE to assess canopy chlorophyll content and plant health using the Red Edge band.
}
