% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ndre}
\alias{d4h_ndre}
\title{Normalized Difference Red Edge Index (NDRE)}
\usage{
d4h_ndre(nir, red_edge)
}
\arguments{
\item{nir}{SpatRaster. Near-infrared band.}

\item{red_edge}{SpatRaster. Red Edge band.}
}
\value{
A SpatRaster containing NDRE values.
}
\description{
Calculates NDRE using the Red Edge band.
}
