% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{d4h_ndvi}
\alias{d4h_ndvi}
\title{Calcular NDVI a partir de capas o rutas (con opción a grilla zonal)}
\usage{
d4h_ndvi(nir, red, tamano_celda = NULL, square = TRUE, escala = NULL)
}
\description{
Calcular NDVI a partir de capas o rutas (con opción a grilla zonal)
}
