% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{d4h_ndvi_general}
\alias{d4h_ndvi_general}
\title{Cálculo general continuo de NDVI para toda el área (sin grilla)}
\usage{
d4h_ndvi_general(nir, red, escala = NULL)
}
\description{
Cálculo general continuo de NDVI para toda el área (sin grilla)
}
