% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ndwi}
\alias{d4h_ndwi}
\title{Normalized Difference Water Index (NDWI)}
\usage{
d4h_ndwi(green, nir)
}
\arguments{
\item{green}{SpatRaster. Green band.}

\item{nir}{SpatRaster. Near-infrared band.}
}
\value{
A SpatRaster containing NDWI values.
}
\description{
Calculates NDWI to identify water bodies and surface moisture.
}
