% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_ndwi}
\alias{d4h_ndwi}
\title{Normalized Difference Water Index (NDWI)}
\usage{
d4h_ndwi(green, nir)
}
\arguments{
\item{green}{SpatRaster or character. Green band or path to file.}

\item{nir}{SpatRaster or character. Near-infrared band or path to file.}
}
\value{
A SpatRaster containing NDWI values in \code{[-1, 1]} with background masked to NA.
}
\description{
Calculates NDWI to detect open surface water bodies and moisture accumulation.
}
