% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_d4h.R
\name{d4h_plot}
\alias{d4h_plot}
\title{Cartographic Visualization with ggplot2}
\usage{
d4h_plot(
  x,
  grid = NULL,
  palette = "viridis",
  limits = NULL,
  title = NULL,
  max_pixels = 5e+05
)
}
\arguments{
\item{x}{SpatRaster, SpatVector, or sf object to plot.}

\item{grid}{Optional SpatVector or sf object to overlay grid borders.}

\item{palette}{Character. Color palette ("viridis", "magma", "terrain" or custom colors vector).}

\item{limits}{Numeric vector c(min, max) to fix the color scale limits.}

\item{title}{Character. Plot title.}

\item{max_pixels}{Integer. Maximum number of pixels to sample for fast rendering (default: 500000).}
}
\value{
A ggplot object.
}
\description{
Plots continuous raster layers (SpatRaster) or zonal grids (SpatVector / sf) with ggplot2.
}
