% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_d4h.R
\name{d4h_plot}
\alias{d4h_plot}
\title{Visualización cartográfica con ggplot2}
\usage{
d4h_plot(
  x,
  grilla = NULL,
  paleta = "viridis",
  limites = NULL,
  titulo = NULL,
  max_pixels = 5e+05
)
}
\arguments{
\item{x}{Objeto SpatRaster o SpatVector a graficar.}

\item{grilla}{Objeto SpatVector o sf opcional para superponer bordes de malla.}

\item{paleta}{Opciones: "viridis", "magma", "terrain" o vector de colores.}

\item{limites}{Vector numérico c(min, max) para fijar la escala de color.}

\item{titulo}{Título del gráfico.}

\item{max_pixels}{Número máximo de píxeles a muestrear para visualización ágil (default: 500000).}
}
\value{
Objeto ggplot.
}
\description{
Grafica capas ráster (SpatRaster) o grillas zonales (SpatVector / sf) con ggplot2.
Omite el área fuera del vuelo del dron y permite superponer grillas.
}
