% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{d4h_savi}
\alias{d4h_savi}
\title{Calcular SAVI a partir de capas o rutas (con opción a grilla zonal)}
\usage{
d4h_savi(nir, red, L = 0.5, tamano_celda = NULL, square = TRUE, escala = NULL)
}
\description{
Calcular SAVI a partir de capas o rutas (con opción a grilla zonal)
}
