% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_terrain.R
\name{d4h_slope}
\alias{d4h_slope}
\title{Topographic Slope}
\usage{
d4h_slope(dem_raster, unit = "degrees")
}
\arguments{
\item{dem_raster}{SpatRaster. Digital Elevation Model.}

\item{unit}{Character. Output unit ("degrees" or "radians").}
}
\value{
A SpatRaster representing terrain slope.
}
\description{
Derives terrain slope from a Digital Elevation Model (DEM).
}
