% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_terrain.R
\name{d4h_twi}
\alias{d4h_twi}
\title{Topographic Wetness Index (TWI)}
\usage{
d4h_twi(dem_raster, eps = 0.001)
}
\arguments{
\item{dem_raster}{SpatRaster or character. Input DEM/DSM or path to file.}

\item{eps}{Numeric. Small epsilon to prevent division by zero in flat areas (default: 0.001).}
}
\value{
A SpatRaster containing TWI values with background masked to NA.
}
\description{
Calculates Topographic Wetness Index to model micro-scale hydrological accumulation zones.
}
