% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_terrain.R
\name{d4h_twi}
\alias{d4h_twi}
\title{Topographic Wetness Index (TWI)}
\usage{
d4h_twi(dem_raster, eps = 0.001)
}
\arguments{
\item{dem_raster}{SpatRaster. Digital Elevation Model.}

\item{eps}{Numeric. Small epsilon to prevent division by zero.}
}
\value{
A SpatRaster containing TWI values.
}
\description{
Calculates TWI to identify potential water accumulation areas based on terrain.
}
