% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_spectral.R
\name{d4h_vari}
\alias{d4h_vari}
\title{Visible Atmospherically Resistant Index (VARI)}
\usage{
d4h_vari(green, red, blue, eps = 0.001)
}
\arguments{
\item{green}{SpatRaster or character. Green band or path to file.}

\item{red}{SpatRaster or character. Red band or path to file.}

\item{blue}{SpatRaster or character. Blue band (or Red Edge / NIR if Blue unavailable) or path to file.}

\item{eps}{Numeric. Small epsilon to prevent zero division (default: 0.001).}
}
\value{
A SpatRaster containing VARI values in \code{[-1, 1]} with background masked to NA.
}
\description{
Computes VARI to assess vegetation fraction while minimizing atmospheric and illumination sensitivity.
}
